Install the version of Ruby I expect
Because I use chruby and ruby-install, not rbenv, Norm's
ruby-autoinstall stuff doesn't work for me. That's fine: this is enough
for what I need, and installing a new ruby is easy enough.

Unfortunately, because of a bug in Norm's manual installation of
Homebrew, we need to fix the permissions for the symlink (note the `-h`
to indicate to change permissions for the link itself) to `brew`.

We also want, once the Ruby version is installed, to install bundler.