Add homedir et al to Finder sidebar favourites
This used to be possible using a built-in tool that ships with macOS
(`sfltool`), but that [seems to have been neutered in 10.13 with no
plans to reinstate
it](https://www.jamf.com/jamf-nation/discussions/20218/the-slftool-thread#responseChild158766).

The only approach that seems to still be working is to use
[FinderSidebarEditor](https://github.com/robperc/FinderSidebarEditor), so
that's what I'm doing.

`sidebar.add()` always pushes the listed items to the top of the list, so
we add them in reverse order.