More reliably check for existing Ruby install
If ~/.ruby-version doesn't specify an exact Ruby version, this would
always try to (re)install the latest version of Ruby. Using
`ruby-install`'s `--no-reinstall` flag instead means it doesn't attempt
to reinstall a version of Ruby that we already have, even if we don't
specify precise version numbers.

This also ensures that *every* installed version of Ruby has Bundler,
which is what we usually want. It might upgrade Bundler too
aggressively, but Bundler being *too* new has never been an issue for
me in the past, so this is probably fine.