Install and configure Backblaze for backups
I've started using [Backblaze](https://www.backblaze.com) for off-site
backups; this installs and configures the client on all my Machines.

I did try to avoid having to set the two environment variables by using
[`security`](https://developer.apple.com/legacy/library/documentation/Darwin/Reference/ManPages/man1/security.1.html),
but [it would
appear](https://apple.stackexchange.com/questions/238296/dump-icloud-keychain-in-terminal/316589#316589)
that it hasn't been updated to access iCloud Keychain (which [uses a
different format than traditional
keychains](https://apple.stackexchange.com/questions/297691/is-it-possible-to-access-and-modify-icloud-keychain-from-command-line-or-automa/316220#316220)).

Given all of that, it seems like environment variables are the only way :(