Replace suited with a small shell script
`suited` was very good for what I wanted at the time, but I keep hitting
its limitations. Norm isn't maintaining it anymore, and it's… a lot.

This replaces it with a much smaller and more targeted-at-my-needs
script. The script can be run directly from the web, and downloads
sub-scripts if so.

Because we're sourcing the sub-scripts, they also get access to our
variables, meaning they can switch behaviours if needs be.