Allow environment-specific configuration
For dev/server systems (which are currently the only types I have), I
want specific sets of things.

This allows me to choose between them at runtime, rather than depending
on having a particular hostname (which I can't necessarily control,
meaning I might have to futz with the names in the repo).

This also removes the extra, unnecessary `github_root` variable in
favour of just using the envvar of the same name.